+++ src/OpenTelemetry.Exporter.Console/ConsoleExporterMetricsExtensions.cs
@@ -39,12 +39,13 @@ namespace OpenTelemetry.Metrics
 
             var exporter = new ConsoleMetricExporter(options);
 
-            if (options.MetricExportIntervalMilliseconds == Timeout.Infinite)
-            {
-                return builder.AddMetricReader(new BaseExportingMetricReader(exporter));
-            }
+            var reader = options.MetricExportIntervalMilliseconds == Timeout.Infinite
+                ? new BaseExportingMetricReader(exporter)
+                : new PeriodicExportingMetricReader(exporter, options.MetricExportIntervalMilliseconds);
 
-            return builder.AddMetricReader(new PeriodicExportingMetricReader(exporter, options.MetricExportIntervalMilliseconds));
+            reader.PreferredAggregationTemporality = options.AggregationTemporality;
+
+            return builder.AddReader(reader);
         }
     }
 }