+++ src/OpenTelemetry.Extensions.Hosting/Implementation/TracerProviderBuilderHosting.cs
@@ -47,10 +47,16 @@ namespace OpenTelemetry.Trace
 
         public TracerProvider Build(IServiceProvider serviceProvider)
         {
-            int i = 0;
-            while (i < this.configurationActions.Count)
+            if (serviceProvider == null)
             {
-                this.configurationActions[i++](serviceProvider, this);
+                throw new ArgumentNullException(nameof(serviceProvider));
+            }
+
+            // Note: Not using a foreach loop because additional actions can be
+            // added during each call.
+            for (int i = 0; i < this.configurationActions.Count; i++)
+            {
+                this.configurationActions[i](serviceProvider, this);
             }
 
             return this.Build();