+++ src/OpenTelemetry.Instrumentation.Http/MeterProviderBuilderExtensions.cs
@@ -44,7 +44,7 @@ namespace OpenTelemetry.Metrics
             //   RecordException - probably doesn't make sense for metric instrumentation
 
             var instrumentation = new HttpClientMetrics();
-            builder.AddSource(HttpClientMetrics.InstrumentationName);
+            builder.AddMeter(HttpClientMetrics.InstrumentationName);
             return builder.AddInstrumentation(() => instrumentation);
         }
     }