+++ src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -17,68 +17,135 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using System.Threading;
 
 namespace OpenTelemetry.Metrics
 {
-    internal class AggregatorStore
+    internal sealed class AggregatorStore
     {
         internal const int MaxMetricPoints = 2000;
         private static readonly ObjectArrayEqualityComparer ObjectArrayComparer = new ObjectArrayEqualityComparer();
         private readonly object lockZeroTags = new object();
+        private readonly HashSet<string> tagKeysInteresting;
+        private readonly int tagsKeysInterestingCount;
 
         // Two-Level lookup. TagKeys x [ TagValues x Metrics ]
         private readonly ConcurrentDictionary<string[], ConcurrentDictionary<object[], int>> keyValue2MetricAggs =
             new ConcurrentDictionary<string[], ConcurrentDictionary<object[], int>>(new StringArrayEqualityComparer());
 
-        private AggregationTemporality temporality;
-        private MetricPoint[] metrics;
+        private readonly AggregationTemporality temporality;
+        private readonly bool outputDelta;
+        private readonly MetricPoint[] metricPoints;
+        private readonly AggregationType aggType;
+        private readonly double[] histogramBounds;
+        private readonly UpdateLongDelegate updateLongCallback;
+        private readonly UpdateDoubleDelegate updateDoubleCallback;
         private int metricPointIndex = 0;
         private bool zeroTagMetricPointInitialized;
-        private AggregationType aggType;
         private DateTimeOffset startTimeExclusive;
         private DateTimeOffset endTimeInclusive;
 
-        internal AggregatorStore(AggregationType aggType, AggregationTemporality temporality)
+        internal AggregatorStore(
+            AggregationType aggType,
+            AggregationTemporality temporality,
+            double[] histogramBounds,
+            string[] tagKeysInteresting = null)
         {
-            this.metrics = new MetricPoint[MaxMetricPoints];
+            this.metricPoints = new MetricPoint[MaxMetricPoints];
             this.aggType = aggType;
             this.temporality = temporality;
+            this.outputDelta = temporality == AggregationTemporality.Delta ? true : false;
+            this.histogramBounds = histogramBounds;
             this.startTimeExclusive = DateTimeOffset.UtcNow;
-        }
-
-        internal int FindMetricAggregators(ReadOnlySpan<KeyValuePair<string, object>> tags)
-        {
-            int len = tags.Length;
-            if (len == 0)
+            if (tagKeysInteresting == null)
             {
-                if (!this.zeroTagMetricPointInitialized)
+                this.updateLongCallback = this.UpdateLong;
+                this.updateDoubleCallback = this.UpdateDouble;
+            }
+            else
+            {
+                this.updateLongCallback = this.UpdateLongCustomTags;
+                this.updateDoubleCallback = this.UpdateDoubleCustomTags;
+                var hs = new HashSet<string>(StringComparer.Ordinal);
+                foreach (var key in tagKeysInteresting)
                 {
-                    lock (this.lockZeroTags)
-                    {
-                        if (!this.zeroTagMetricPointInitialized)
-                        {
-                            var dt = DateTimeOffset.UtcNow;
-                            this.metrics[0] = new MetricPoint(this.aggType, dt, null, null);
-                            this.zeroTagMetricPointInitialized = true;
-                        }
-                    }
+                    hs.Add(key);
                 }
 
-                return 0;
+                this.tagKeysInteresting = hs;
+                this.tagsKeysInterestingCount = hs.Count;
             }
+        }
 
-            var storage = ThreadStaticStorage.GetStorage();
+        private delegate void UpdateLongDelegate(long value, ReadOnlySpan<KeyValuePair<string, object>> tags);
 
-            storage.SplitToKeysAndValues(tags, out var tagKey, out var tagValue);
+        private delegate void UpdateDoubleDelegate(double value, ReadOnlySpan<KeyValuePair<string, object>> tags);
 
-            if (len > 1)
+        internal void Update(long value, ReadOnlySpan<KeyValuePair<string, object>> tags)
+        {
+            this.updateLongCallback(value, tags);
+        }
+
+        internal void Update(double value, ReadOnlySpan<KeyValuePair<string, object>> tags)
+        {
+            this.updateDoubleCallback(value, tags);
+        }
+
+        internal void SnapShot()
+        {
+            var indexSnapShot = Math.Min(this.metricPointIndex, MaxMetricPoints - 1);
+
+            for (int i = 0; i <= indexSnapShot; i++)
             {
-                Array.Sort<string, object>(tagKey, tagValue);
+                ref var metricPoint = ref this.metricPoints[i];
+                if (metricPoint.StartTime == default)
+                {
+                    continue;
+                }
+
+                metricPoint.TakeSnapShot(this.outputDelta);
             }
 
-            int aggregatorIndex;
+            if (this.temporality == AggregationTemporality.Delta)
+            {
+                if (this.endTimeInclusive != default)
+                {
+                    this.startTimeExclusive = this.endTimeInclusive;
+                }
+            }
 
+            DateTimeOffset dt = DateTimeOffset.UtcNow;
+            this.endTimeInclusive = dt;
+        }
+
+        internal BatchMetricPoint GetMetricPoints()
+        {
+            var indexSnapShot = Math.Min(this.metricPointIndex, MaxMetricPoints - 1);
+            return new BatchMetricPoint(this.metricPoints, indexSnapShot + 1, this.startTimeExclusive, this.endTimeInclusive);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private void InitializeZeroTagPointIfNotInitialized()
+        {
+            if (!this.zeroTagMetricPointInitialized)
+            {
+                lock (this.lockZeroTags)
+                {
+                    if (!this.zeroTagMetricPointInitialized)
+                    {
+                        var dt = DateTimeOffset.UtcNow;
+                        this.metricPoints[0] = new MetricPoint(this.aggType, dt, null, null, this.histogramBounds);
+                        this.zeroTagMetricPointInitialized = true;
+                    }
+                }
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private int LookupAggregatorStore(string[] tagKey, object[] tagValue, int length)
+        {
+            int aggregatorIndex;
             string[] seqKey = null;
 
             // GetOrAdd by TagKey at 1st Level of 2-level dictionary structure.
@@ -86,7 +153,7 @@ namespace OpenTelemetry.Metrics
             if (!this.keyValue2MetricAggs.TryGetValue(tagKey, out var value2metrics))
             {
                 // Note: We are using storage from ThreadStatic, so need to make a deep copy for Dictionary storage.
-                seqKey = new string[len];
+                seqKey = new string[length];
                 tagKey.CopyTo(seqKey, 0);
 
                 value2metrics = new ConcurrentDictionary<object[], int>(ObjectArrayComparer);
@@ -128,16 +195,16 @@ namespace OpenTelemetry.Metrics
                         // Note: We are using storage from ThreadStatic, so need to make a deep copy for Dictionary storage.
                         if (seqKey == null)
                         {
-                            seqKey = new string[len];
+                            seqKey = new string[length];
                             tagKey.CopyTo(seqKey, 0);
                         }
 
-                        var seqVal = new object[len];
+                        var seqVal = new object[length];
                         tagValue.CopyTo(seqVal, 0);
 
-                        ref var metricPoint = ref this.metrics[aggregatorIndex];
+                        ref var metricPoint = ref this.metricPoints[aggregatorIndex];
                         var dt = DateTimeOffset.UtcNow;
-                        metricPoint = new MetricPoint(this.aggType, dt, seqKey, seqVal);
+                        metricPoint = new MetricPoint(this.aggType, dt, seqKey, seqVal, this.histogramBounds);
 
                         // Add to dictionary *after* initializing MetricPoint
                         // as other threads can start writing to the
@@ -150,18 +217,18 @@ namespace OpenTelemetry.Metrics
             return aggregatorIndex;
         }
 
-        internal void UpdateLong(long value, ReadOnlySpan<KeyValuePair<string, object>> tags)
+        private void UpdateLong(long value, ReadOnlySpan<KeyValuePair<string, object>> tags)
         {
             try
             {
-                var index = this.FindMetricAggregators(tags);
+                var index = this.FindMetricAggregatorsDefault(tags);
                 if (index < 0)
                 {
                     // TODO: Measurement dropped due to MemoryPoint cap hit.
                     return;
                 }
 
-                this.metrics[index].Update(value);
+                this.metricPoints[index].Update(value);
             }
             catch (Exception)
             {
@@ -169,18 +236,18 @@ namespace OpenTelemetry.Metrics
             }
         }
 
-        internal void UpdateDouble(double value, ReadOnlySpan<KeyValuePair<string, object>> tags)
+        private void UpdateLongCustomTags(long value, ReadOnlySpan<KeyValuePair<string, object>> tags)
         {
             try
             {
-                var index = this.FindMetricAggregators(tags);
+                var index = this.FindMetricAggregatorsCustomTag(tags);
                 if (index < 0)
                 {
                     // TODO: Measurement dropped due to MemoryPoint cap hit.
                     return;
                 }
 
-                this.metrics[index].Update(value);
+                this.metricPoints[index].Update(value);
             }
             catch (Exception)
             {
@@ -188,37 +255,96 @@ namespace OpenTelemetry.Metrics
             }
         }
 
-        internal void SnapShot()
+        private void UpdateDouble(double value, ReadOnlySpan<KeyValuePair<string, object>> tags)
         {
-            var indexSnapShot = Math.Min(this.metricPointIndex, MaxMetricPoints - 1);
-
-            for (int i = 0; i <= indexSnapShot; i++)
+            try
             {
-                ref var metricPoint = ref this.metrics[i];
-                if (metricPoint.StartTime == default)
+                var index = this.FindMetricAggregatorsDefault(tags);
+                if (index < 0)
                 {
-                    continue;
+                    // TODO: Measurement dropped due to MemoryPoint cap hit.
+                    return;
                 }
 
-                metricPoint.TakeSnapShot(this.temporality == AggregationTemporality.Delta ? true : false);
+                this.metricPoints[index].Update(value);
             }
-
-            if (this.temporality == AggregationTemporality.Delta)
+            catch (Exception)
             {
-                if (this.endTimeInclusive != default)
-                {
-                    this.startTimeExclusive = this.endTimeInclusive;
-                }
+                // TODO: Measurement dropped due to internal exception.
             }
-
-            DateTimeOffset dt = DateTimeOffset.UtcNow;
-            this.endTimeInclusive = dt;
         }
 
-        internal BatchMetricPoint GetMetricPoints()
+        private void UpdateDoubleCustomTags(double value, ReadOnlySpan<KeyValuePair<string, object>> tags)
         {
-            var indexSnapShot = Math.Min(this.metricPointIndex, MaxMetricPoints - 1);
-            return new BatchMetricPoint(this.metrics, indexSnapShot + 1, this.startTimeExclusive, this.endTimeInclusive);
+            try
+            {
+                var index = this.FindMetricAggregatorsCustomTag(tags);
+                if (index < 0)
+                {
+                    // TODO: Measurement dropped due to MemoryPoint cap hit.
+                    return;
+                }
+
+                this.metricPoints[index].Update(value);
+            }
+            catch (Exception)
+            {
+                // TODO: Measurement dropped due to internal exception.
+            }
+        }
+
+        private int FindMetricAggregatorsDefault(ReadOnlySpan<KeyValuePair<string, object>> tags)
+        {
+            int tagLength = tags.Length;
+            if (tagLength == 0)
+            {
+                this.InitializeZeroTagPointIfNotInitialized();
+                return 0;
+            }
+
+            var storage = ThreadStaticStorage.GetStorage();
+
+            storage.SplitToKeysAndValues(tags, tagLength, out var tagKey, out var tagValue);
+
+            if (tagLength > 1)
+            {
+                Array.Sort<string, object>(tagKey, tagValue);
+            }
+
+            return this.LookupAggregatorStore(tagKey, tagValue, tagLength);
+        }
+
+        private int FindMetricAggregatorsCustomTag(ReadOnlySpan<KeyValuePair<string, object>> tags)
+        {
+            int tagLength = tags.Length;
+            if (tagLength == 0 || this.tagsKeysInterestingCount == 0)
+            {
+                this.InitializeZeroTagPointIfNotInitialized();
+                return 0;
+            }
+
+            // TODO: Get only interesting tags
+            // from the incoming tags
+
+            var storage = ThreadStaticStorage.GetStorage();
+
+            storage.SplitToKeysAndValues(tags, tagLength, this.tagKeysInteresting, out var tagKey, out var tagValue, out var actualLength);
+
+            // Actual number of tags depend on how many
+            // of the incoming tags has user opted to
+            // select.
+            if (actualLength == 0)
+            {
+                this.InitializeZeroTagPointIfNotInitialized();
+                return 0;
+            }
+
+            if (actualLength > 1)
+            {
+                Array.Sort<string, object>(tagKey, tagValue);
+            }
+
+            return this.LookupAggregatorStore(tagKey, tagValue, actualLength);
         }
     }
 }