+++ src/OpenTelemetry/Metrics/CompositeMetricReader.cs
@@ -62,6 +62,8 @@ namespace OpenTelemetry.Metrics
             return this;
         }
 
+        public Enumerator GetEnumerator() => new Enumerator(this.head);
+
         /// <inheritdoc/>
         protected override bool ProcessMetrics(Batch<Metric> metrics, int timeoutMilliseconds)
         {
@@ -154,7 +156,32 @@ namespace OpenTelemetry.Metrics
             this.disposed = true;
         }
 
-        private class DoublyLinkedListNode
+        public struct Enumerator
+        {
+            private DoublyLinkedListNode node;
+
+            internal Enumerator(DoublyLinkedListNode node)
+            {
+                this.node = node;
+                this.Current = null;
+            }
+
+            public MetricReader Current { get; private set; }
+
+            public bool MoveNext()
+            {
+                if (this.node != null)
+                {
+                    this.Current = this.node.Value;
+                    this.node = this.node.Next;
+                    return true;
+                }
+
+                return false;
+            }
+        }
+
+        internal class DoublyLinkedListNode
         {
             public readonly MetricReader Value;