+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpTraceExporterTests.cs
@@ -19,7 +19,9 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading;
+using Moq;
 using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation;
+using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Tests;
 using OpenTelemetry.Trace;
@@ -68,10 +70,6 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
                 new ActivitySource("odd", "1.3.5"),
             };
 
-            using var exporter = new OtlpTraceExporter(
-                new OtlpExporterOptions(),
-                new NoopTraceServiceClient());
-
             var resourceBuilder = ResourceBuilder.CreateEmpty();
             if (includeServiceNameInResource)
             {
@@ -89,8 +87,6 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
                 .AddSource(sources[1].Name);
 
             using var openTelemetrySdk = builder.Build();
-
-            exporter.ParentProvider = openTelemetrySdk;
 
             var processor = new BatchActivityExportProcessor(new TestExporter<Activity>(RunTest));
             const int numOfSpans = 10;
@@ -112,18 +108,18 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             {
                 var request = new OtlpCollector.ExportTraceServiceRequest();
 
-                request.AddBatch(exporter.ProcessResource, batch);
+                request.AddBatch(resourceBuilder.Build().ToOtlpResource(), batch);
 
                 Assert.Single(request.ResourceSpans);
                 var oltpResource = request.ResourceSpans.First().Resource;
                 if (includeServiceNameInResource)
                 {
-                    Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == Resources.ResourceSemanticConventions.AttributeServiceName && kvp.Value.StringValue == "service-name");
-                    Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == Resources.ResourceSemanticConventions.AttributeServiceNamespace && kvp.Value.StringValue == "ns1");
+                    Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == ResourceSemanticConventions.AttributeServiceName && kvp.Value.StringValue == "service-name");
+                    Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == ResourceSemanticConventions.AttributeServiceNamespace && kvp.Value.StringValue == "ns1");
                 }
                 else
                 {
-                    Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == Resources.ResourceSemanticConventions.AttributeServiceName && kvp.Value.ToString().Contains("unknown_service:"));
+                    Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == ResourceSemanticConventions.AttributeServiceName && kvp.Value.ToString().Contains("unknown_service:"));
                 }
 
                 foreach (var instrumentationLibrarySpans in request.ResourceSpans.First().InstrumentationLibrarySpans)
@@ -336,6 +332,18 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 
             Assert.True(startCalled);
             Assert.True(endCalled);
+        }
+
+        [Fact]
+        public void Shutdown_ClientShutdownIsCalled()
+        {
+            var exportClientMock = new Mock<IExportClient<OtlpCollector.ExportTraceServiceRequest>>();
+
+            var exporter = new OtlpTraceExporter(new OtlpExporterOptions(), exportClientMock.Object);
+
+            var result = exporter.Shutdown();
+
+            exportClientMock.Verify(m => m.Shutdown(It.IsAny<int>()), Times.Once());
         }
 
         private class NoopTraceServiceClient : OtlpCollector.TraceService.ITraceServiceClient