+++ test/OpenTelemetry.Tests/Trace/TracerProviderSdkTest.cs
@@ -949,6 +949,81 @@ namespace OpenTelemetry.Trace.Tests
             Assert.True(testActivityProcessor.ForceFlushCalled);
         }
 
+        [Fact]
+        public void SdkSamplesAndProcessesLegacySourceWhenAddLegacySourceIsCalledWithWildcardValue()
+        {
+            var sampledActivities = new List<string>();
+            var sampler = new TestSampler
+            {
+                SamplingAction =
+                (samplingParameters) =>
+                {
+                    sampledActivities.Add(samplingParameters.Name);
+                    return new SamplingResult(SamplingDecision.RecordAndSample);
+                },
+            };
+
+            using TestActivityProcessor testActivityProcessor = new TestActivityProcessor();
+
+            var onStartProcessedActivities = new List<string>();
+            var onStopProcessedActivities = new List<string>();
+            testActivityProcessor.StartAction =
+                (a) =>
+                {
+                    Assert.Contains(a.OperationName, sampledActivities);
+                    Assert.False(Sdk.SuppressInstrumentation);
+                    Assert.True(a.IsAllDataRequested); // If Proccessor.OnStart is called, activity's IsAllDataRequested is set to true
+                    onStartProcessedActivities.Add(a.OperationName);
+                };
+
+            testActivityProcessor.EndAction =
+                (a) =>
+                {
+                    Assert.False(Sdk.SuppressInstrumentation);
+                    Assert.True(a.IsAllDataRequested); // If Processor.OnEnd is called, activity's IsAllDataRequested is set to true
+                    onStopProcessedActivities.Add(a.OperationName);
+                };
+
+            var legacySourceNamespaces = new[] { "LegacyNamespace.*", "Namespace.*.Operation" };
+            using var activitySource = new ActivitySource(ActivitySourceName);
+
+            // AddLegacyOperationName chained to TracerProviderBuilder
+            using var tracerProvider = Sdk.CreateTracerProviderBuilder()
+                        .SetSampler(sampler)
+                        .AddProcessor(testActivityProcessor)
+                        .AddLegacySource(legacySourceNamespaces[0])
+                        .AddLegacySource(legacySourceNamespaces[1])
+                        .AddSource(ActivitySourceName)
+                        .Build();
+
+            foreach (var ns in legacySourceNamespaces)
+            {
+                var startOpName = ns.Replace("*", "Start");
+                Activity startOperation = new Activity(startOpName);
+                startOperation.Start();
+                startOperation.Stop();
+
+                Assert.Contains(startOpName, onStartProcessedActivities); // Processor.OnStart is called since we added a legacy OperationName
+                Assert.Contains(startOpName, onStopProcessedActivities);  // Processor.OnEnd is called since we added a legacy OperationName
+
+                var stopOpName = ns.Replace("*", "Stop");
+                Activity stopOperation = new Activity(stopOpName);
+                stopOperation.Start();
+                stopOperation.Stop();
+
+                Assert.Contains(stopOpName, onStartProcessedActivities); // Processor.OnStart is called since we added a legacy OperationName
+                Assert.Contains(stopOpName, onStopProcessedActivities);  // Processor.OnEnd is called since we added a legacy OperationName
+            }
+
+            // Ensure we can still process "normal" activities when in legacy wildcard mode.
+            Activity nonLegacyActivity = activitySource.StartActivity("TestActivity");
+            nonLegacyActivity.Start();
+            nonLegacyActivity.Stop();
+
+            Assert.Contains(nonLegacyActivity.OperationName, onStartProcessedActivities); // Processor.OnStart is called since we added a legacy OperationName
+            Assert.Contains(nonLegacyActivity.OperationName, onStopProcessedActivities);  // Processor.OnEnd is called since we added a legacy OperationName
+        }
+
         public void Dispose()
         {
             GC.SuppressFinalize(this);